Report accuracy for False and True separately
This commit makes the classifier to report the accuracy for False label
and True label separately to find overfitting.
The elements of the attribute are mainly False because the program uses
only a few functions. Due to this label imbalance, the common
binary_accuracy is not suitable as a metric of this training. We cannot
find overfitting to False label by using the only binary_accuracy.